models/SVD/requirements.txt
torch>=2.1.0
torchvision>=0.16.0

diffusers>=0.27.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

Pillow
imageio[ffmpeg]
numpy

jupyter
matplotlib
